password generator
This is a program that is used to generate a strong and unique password for user depending on the length they require. A password generator is a software tool that creates random or customized passwords for users. It helps users create stronger passwords that provide greater security for a given type of access.