numpy
In this project we havecreate a code such that the system will randomly pick any numberand gives the status of it.
That is If it’s a positive number it shows positive number orIf it’s negative number it shows negative number. If it’s even number it shows even number or If it’s odd number it shows odd number.